requirements.txt
streamlit
pandas
qrcode
opencv-python-headless
Pillow
pyzbar
